rename funcs to utils
