+++ cogs5e/dice.py
@@ -6,7 +6,7 @@ import discord
 from discord.ext import commands
 
 from aliasing import helpers
-from cogs5e.funcs import attackutils, checkutils, targetutils
+from cogs5e.utils import attackutils, checkutils, targetutils
 from cogs5e.models.errors import NoSelectionElements
 from cogsmisc.stats import Stats
 from gamedata import Monster
