+++ cogs5e/gametrack.py
@@ -13,7 +13,7 @@ import discord
 from discord.ext import commands
 
 from aliasing import helpers
-from cogs5e.funcs import targetutils
+from cogs5e.utils import targetutils
 from cogs5e.models.character import Character, CustomCounter
 from cogs5e.models.embeds import EmbedWithCharacter
 from cogs5e.models.errors import ConsumableException, CounterOutOfBounds, InvalidArgument, NoSelectionElements
