+++ cogs5e/utils/checkutils.py
