+++ cogs5e/utils/gamelogutils.py
